simplify prepare command
