+++ tipocket-ctl/tpctl/__main__.py
@@ -1,10 +1,10 @@
 from tpctl.app import cli
 from tpctl.debug import debug
-from tpctl.prepare import prepare
+from tpctl.deploy import deploy
 
 
 def main():
-    cli.add_command(prepare)
+    cli.add_command(deploy)
     cli.add_command(debug)
     cli()
